# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API REST que busca e cadastra pessoas no banco de dados. Criada com Laravel e JWT.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Faz login e retorna o token necessário para o uso da API",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@mail.com",
                                    "password": "147258369"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "access_token": "xxx.xxx.xxx",
                                        "token_type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Retorna dados sobre o usuário logado",
                "operationId": "f0b15943f55143d36f606a2dbd1d2366",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "id": 1,
                                        "name": "Administrador",
                                        "email": "admin@mail.com",
                                        "email_verified_at": null,
                                        "created_at": "2023-06-18T18:27:36.000000Z",
                                        "updated_at": "2023-06-18T18:27:36.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Faz logout",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Atualiza o token do usuário logado",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "access_token": "xxx.xxx.xxx",
                                        "token_type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get/{id}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Busca uma pessoa do banco de dados",
                "operationId": "a1242ab196b5e579b6b627906cc20ad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "id": 1,
                                        "name": "João Pedro"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getAll": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Busca todas as pessoas registradas no banco de dados",
                "operationId": "6adb876cd4fe6eed93757c541f145a09",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "name": "João Pedro"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Maria Luiza"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/create": {
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Adiciona uma nova pessoa no banco de dados",
                "operationId": "09f070d0c00146b340f50c192edb0c53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome da pessoa",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update": {
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Altera o registro de uma pessoa no banco de dados",
                "operationId": "a479b88efc9064832b944f2a9896c80e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Id da pessoa",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Novo nome da pessoa",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete": {
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Apaga o registro de uma pessoa do banco de dados",
                "operationId": "6d1255f4a31a00a700298780a140c711",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Id da pessoa",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Sem autenticação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}